Added files via upload
Python Program that has three parts. 
The first part takes in user input for three exam scores, and outputs an average. 
The second part takes in user input for a first name, middle name, and last name, and outputs the initials. 
The third part takes in user input for an amount of seconds, and outputs the time in Hours:Minutes:Seconds format.